feat: Add platform overview components and integrate into campaign analytics
- Created new components for platform overview (mobile and desktop) in campaign-analytics.
- Updated EngagementRateHeatmap and TopCreatorsLineChart for responsive design adjustments.
- Refactored CampaignAnalytics to utilize new platform overview components for better layout management.
- Enhanced styling for various elements to improve mobile responsiveness and visual consistency.
- Minor adjustments to typography and spacing for better readability across different screen sizes.
