+++ src/components/campaign-analytics/index.js
@@ -1,0 +1,3 @@
+// Export all platform overview components
+export { default as PlatformOverviewMobile } from './PlatformOverviewMobile';
+export { default as PlatformOverviewDesktop } from './PlatformOverviewDesktop';
